refactor: 관리자 빙고 편집기 컴포넌트 구조 분리
